requirements.txt
# Core dependencies
fastmcp>=0.9.0
qdrant-client>=1.7.0
fastembed>=0.2.0
python-dotenv>=1.0.0

# LLM providers (optional - install based on usage)
openai>=1.0.0
anthropic>=0.20.0
httpx>=0.25.0  # For local LLM APIs
google-genai>=0.5.0

# Document processing
PyPDF2>=3.0.0
python-docx>=1.1.0
beautifulsoup4>=4.12.0
pandas>=2.0.0  # For CSV/Excel processing
openpyxl>=3.1.0  # For Excel files

# Development and testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0
pytest-mock>=3.12.0

# Utilities
pathlib2>=2.3.0  # For Python < 3.4 compatibility (if needed)
typing-extensions>=4.8.0

# Logging and monitoring
structlog>=23.1.0  # Structured logging support

# Security
cryptography>=41.0.0  # For secure operations if needed